# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Wavelet 平台后端 API，提供用户认证、系统配置、任务调度等通用功能。",
        "title": "Wavelet API",
        "contact": {
            "name": "Wavelet",
            "url": "https://github.com/Rain-kl/Wavelet"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/upload": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "支持各种类型的通用文件上传，支持自动文件类型检测、哈希计算与“秒传”去重",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "要上传的文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "业务分类 (例如: avatar, attachment, doc，默认为 generic)",
                        "name": "type",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "额外的 JSON 格式元数据",
                        "name": "metadata",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseAny"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Upload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误或文件受限",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
        "/api/v1/upload/download/batch": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "传入多个文件 ID，后台实时将其打包压缩为 ZIP 流并输出，自动处理文件名重复冲突",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "批量打包下载",
                "parameters": [
                    {
                        "description": "包含文件 ID 数组的请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/upload.batchDownloadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功下载打包后的 ZIP",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "500": {
                        "description": "打包失败",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
        "/api/v1/upload/download/{id}": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "根据文件 ID 获取文件，以附件形式 (Attachment) 强制开启客户端浏览器下载",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "下载单文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功下载文件",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "404": {
                        "description": "文件不存在",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "500": {
                        "description": "服务内部错误",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
        "/api/v1/upload/my": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "分页获取当前登录用户上传的文件，支持文件名关键词、业务类型、扩展名过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "获取我的文件列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（默认 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量（默认 20，最大 100）",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "文件名关键词（模糊匹配）",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "业务分类过滤",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "扩展名过滤",
                        "name": "extension",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseAny"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/upload.listMyFilesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
        "/api/v1/upload/{id}": {
            "delete": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "将文件状态置为 deleted（软删除），不会立即清理底层存储对象",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "删除文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "403": {
                        "description": "无权操作",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    },
                    "404": {
                        "description": "文件不存在",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Upload": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "extension": {
                    "description": "文件后缀名 (不含点，如 png, pdf)",
                    "type": "string"
                },
                "file_name": {
                    "description": "原始文件名 (例如: image.png)",
                    "type": "string"
                },
                "file_path": {
                    "description": "文件相对路径 / S3 Key",
                    "type": "string"
                },
                "file_size": {
                    "description": "文件大小（字节）",
                    "type": "integer"
                },
                "hash": {
                    "description": "文件哈希 (SHA-256/MD5，可用于排重)",
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "metadata": {
                    "description": "业务扩展元数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.UploadMetadata"
                        }
                    ]
                },
                "mime_type": {
                    "description": "媒体类型 (MIME, 如 image/png)",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.UploadStatus"
                        }
                    ]
                },
                "storage_driver": {
                    "description": "存储引擎驱动 (如 local, s3, oss)",
                    "type": "string"
                },
                "type": {
                    "description": "业务标识类型 (如 avatar, doc, attachment)",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
        "model.UploadMetadata": {
            "type": "object",
            "properties": {
                "bucket": {
                    "description": "存储桶名称 (适用于 S3 等)",
                    "type": "string"
                },
                "client_ip": {
                    "description": "上传者 IP",
                    "type": "string"
                },
                "duration": {
                    "description": "音视频时长 (s)",
                    "type": "number"
                },
                "extra": {
                    "description": "其它任意业务自定义元数据",
                    "type": "object",
                    "additionalProperties": {}
                },
                "height": {
                    "description": "图像/视频高度 (px)",
                    "type": "integer"
                },
                "original_mime": {
                    "description": "原始 MIME 类型",
                    "type": "string"
                },
                "user_agent": {
                    "description": "上传者的 UA",
                    "type": "string"
                },
                "width": {
                    "description": "图像/视频宽度 (px)",
                    "type": "integer"
                }
            }
        },
        "model.UploadStatus": {
            "type": "string",
            "enum": [
                "pending",
                "used",
                "deleted"
            ],
            "x-enum-comments": {
                "UploadStatusDeleted": "已删除",
                "UploadStatusPending": "待使用",
                "UploadStatusUsed": "已使用"
            },
            "x-enum-descriptions": [
                "待使用",
                "已使用",
                "已删除"
            ],
            "x-enum-varnames": [
                "UploadStatusPending",
                "UploadStatusUsed",
                "UploadStatusDeleted"
            ]
        },
...
        "upload.batchDownloadRequest": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "upload.listMyFilesResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Upload"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "util.ResponseAny": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string",
                    "example": ""
                }
            }
        }
    },
    "securityDefinitions": {
        "SessionCookie": {
            "type": "apiKey",
            "name": "session",
            "in": "cookie"
        }
    }
}